requirements.txt
alembic==1.0.11
awscli==1.16.219
boto3==1.9.202
botocore==1.12.209
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.3.9
docutils==0.14
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
future==0.17.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Mako==1.1.0
MarkupSafe==1.1.1
mysql-connector-python-rf==2.2.2
pyasn1==0.4.6
pymongo==3.8.0
pytest-runner==5.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
PyYAML==5.1
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.6
sqs-workers==0.4.1
urllib3==1.25.3
Werkzeug==0.15.5
xmltodict==0.12.0
